Added initial axis value to AxisToAxisAction
Beforehand the initial axis value could only be set for
AxisToRelativeAxisActions, for AxisToAxisActions that are mapped to axes
which default to a non-center positon (e.g. triggers) an initial axis
value that can be set to this exact non-center position initializes the
virtual axis to this position on startup. This is relevant if the
AxisToAxisAction is not part of the default mode and the action would
not be initialized to the correct position "automatically".