ToButtonAction, ToKeyAction, ToMouseButtonAction: added mininum activation interval
If set, the minimum activation interval is the minimum time interval for which the
action will be activated, measured from the beginning of the last activation.

This option is meant to ensure that the simulated inputs of a (repeated) action are
long enough to be handled by the target application.