Fix NPE in Input.poll() introduced with 9481fd1
After some reordering of the startup process in the above mentioned
commit, no updated instance of Input with a non-null ControllerInfo
got created, resulting in a NPE in the poll() method.

In order to prevent this problem, a new instance of Input gets created
after GLFW initialization with a ControllerInfo instance.

In addition, multiple defensive checks are introduced.