fix: omit repoll for lock keys
When On-Sreen Keyboard mode is present within a profile and a Lock Key
action is bound to a button, in conjunction with a Switch Mode action,
the mode switch triggers an OSK repoll right after the Lock Key action
is handled.
Before this fix the repoll would inadvertently set the changed flag for
ALL OSK buttons, including the Lock Key buttons. This in turn will cause
Caps and Num Lock to get disabled again instantly, thus nullifying the
effect of the Lock Key action.