feat: redefine meaning of activation interval for single-immediate and single-on-release actions
For repeat actions, the activation interval already directly defined the
minimum and maximum time interval a joystick button, key or mouse button
is held down.

For single-immediate and single-on-release actions, the activation
interval so far was ill-defined, while it was possible to affect the
minimum time interval between repeated activations, in practice
single-immediate and single-on-release actiosn with an activation
interval made no sense.

With this change, the activation interval now directly controls the time
interval a joystick button, key or mouse button is held down in all
cases.
Consequently it is now possible to have single-immediate and
single-on-release actions which perform outputs over a prolonged
time interval.